requirements.txt
# Emergency Detection System - Requirements
# Core Flask and Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Basic Image Processing
Pillow==10.0.1

# Voice Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.11

# Firebase Integration
firebase-admin==6.2.0

# HTTP Requests and API
requests==2.31.0
urllib3==2.0.4

# Environment and Configuration
python-dotenv==1.0.0

# Geolocation
geocoder==1.38.1

# Development and Testing (Optional)
pytest==7.4.2
pytest-flask==1.2.0